BUG: BroadcastPackets sends multiple packets with each program reload
For example, if I load the program once it works as expected.  Stop the
program, everything stops.  Restart the program, looks correct.  Reload
the program, however, and now it sends 2 packets.  Reload again, and
now it sends 3 packets.  They're identical packets, too.  Might be a
bug related to threading?  Going to shelve it for now.